Make runMode middleware and observer
The old runMode middleware was a pain to test. Also it was kind of an
abuse of what middleware is supposed to do- all it really cared about
was whether or not runMode changed.

Switch over to an observerPattern.
